Fix: ensure superuser.allowed is actually set
On tumbleweed in some cases superuser.allowed can be null instead of
false this leads to the user incorrectly being told they do not have
superuser access as superuser.allowed isn't populated yet and is null.
superuser.reload_page_on_change also only fired when the popup for
superuser escalation was opened, not when it was gained.